test fixes

+++ test/process.cpp
@@ -1053,7 +1053,7 @@ TEST_F(sys_call_test, procinfo_processchild_cpuload)
 
 					uint64_t delta = tcpu - lastcpu;
 
-					printf("%d:%d)%ld:%ld)%ld >> %ld\n", (int)callnum, (int)ctid, tinfo->m_pid, tinfo->m_tid, tcpu, delta);
+					printf("%d:%d)%d:%d)%d >> %d\n", (int)callnum, (int)ctid, (int)tinfo->m_pid, (int)tinfo->m_tid, (int)tcpu, (int)delta);
 
 					if(callnum != 0)
 					{
